IGNITE-12887 move test
